Changes to settings control + changes to lang resources (string duplication fix)
