+++ src/AmbientSounds.Uwp/Controls/MoreButton.xaml.cs
@@ -9,7 +9,6 @@ using Windows.UI.ViewManagement;
 using Windows.UI.Xaml;
 using Windows.UI.Xaml.Controls;
 using Windows.UI.Xaml.Media.Animation;
-using AmbientSounds.Views.Dialogs;
 
 // The User Control item template is documented at https://go.microsoft.com/fwlink/?LinkId=234236
