+++ src/AmbientSounds.Uwp/Controls/SettingsControl.xaml.cs
