+++ src/AmbientSounds.Uwp/Strings/cs-CZ/Resources.resw
@@ -247,10 +247,7 @@
     <value>Změnit hlasitost zvuku</value>
   </data>
   <data name="CloseText" xml:space="preserve">
-    <value>Zavřete</value>
-  </data>
-  <data name="MoreButtonSettings.Label" xml:space="preserve">
-    <value>Nastavení</value>
+    <value>Zavřít</value>
   </data>
   <data name="OffTextBlock.Text" xml:space="preserve">
     <value>Zakázán</value>
@@ -261,7 +258,7 @@
   <data name="SettingsTelemetrySwitch.Header" xml:space="preserve">
     <value>Telemetrie</value>
   </data>
-	<data name="SettingsText" xml:space="preserve">
+  <data name="SettingsText" xml:space="preserve">
     <value>Nastavení</value>
   </data>
 </root>
