+++ src/AmbientSounds.Uwp/Strings/es-ES/Resources.resw
@@ -168,9 +168,6 @@
   <data name="MoreButtonRate.Label" xml:space="preserve">
     <value>Valóranos</value>
   </data>
-  <data name="MoreButtonSettings.Label" xml:space="preserve">
-    <value />
-  </data>
   <data name="MoreButtonShare.Label" xml:space="preserve">
     <value>Compartir URL para Ambie</value>
   </data>
