+++ src/AmbientSounds.Uwp/Strings/hu-HU/Resources.resw
@@ -168,9 +168,6 @@
   <data name="MoreButtonRate.Label" xml:space="preserve">
     <value>Ambie értékelése</value>
   </data>
-  <data name="MoreButtonSettings.Label" xml:space="preserve">
-    <value />
-  </data>
   <data name="MoreButtonShare.Label" xml:space="preserve">
     <value>Ambie megosztása linkkel</value>
   </data>
