--- src/AmbientSounds.Uwp/Views/Dialogs/SettingsDialog.xaml
@@ -1,42 +1,0 @@
-﻿<ContentDialog
-    x:Class="AmbientSounds.Views.Dialogs.SettingsDialog"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:AmbientSounds.Views.Dialogs"
-    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    mc:Ignorable="d"
-    Title="Settings"
-    SecondaryButtonText="Close"
-    PrimaryButtonClick="ContentDialog_PrimaryButtonClick"
-    SecondaryButtonClick="ContentDialog_SecondaryButtonClick">
-
-    <ContentDialog.Resources>
-        <Style TargetType="RadioButton">
-            <Setter Property="HorizontalAlignment" Value="Stretch" />
-        </Style>
-
-        <Style TargetType="TextBlock" x:Key="SettingTitleStyle">
-            <Setter Property="FontSize" Value="16" />
-            <Setter Property="FontWeight" Value="Bold" />
-        </Style>
-
-        <Style TargetType="TextBlock" x:Key="ThemeWarningTextStyle">
-            <Setter Property="Margin" Value="0,5,0,0" />
-            <Setter Property="FontSize" Value="12" />
-        </Style>
-    </ContentDialog.Resources>
-
-    <Grid>
-        <StackPanel>
-            <TextBlock x:Uid="AppThemeTextBlock" Text="App theme" Style="{StaticResource SettingTitleStyle}" />
-            <RadioButton x:Uid="DefaultThemeRadio" Content="Use system setting"
-                         Click="{x:Bind ViewModel.DefaultThemeRadioClicked}" IsChecked="{x:Bind ViewModel.DefaultRadioIsChecked}" />
-            <RadioButton x:Uid="DarkThemeRadio" Content="Dark theme"
-                         Click="{x:Bind ViewModel.DarkThemeRadioClicked}" IsChecked="{x:Bind ViewModel.DarkRadioIsChecked}" />
-            <RadioButton x:Uid="LightThemeRadio" Content="Light theme"
-                         Click="{x:Bind ViewModel.LightThemeRadioClicked}" IsChecked="{x:Bind ViewModel.LightRadioIsChecked}" />
-            <TextBlock x:Uid="ThemeWarningTextBlock" Text="Theme will be changed after restart." Name="ThemeWarning" Style="{StaticResource ThemeWarningTextStyle}" />
-        </StackPanel>
-    </Grid>
-</ContentDialog>
